TODOs, object's parameters (area,perimeter,volume,surface), calculate Shape3D Volume and Surface

+++ Thesis/Geometry.h
@@ -92,9 +92,9 @@ inline void Plane2DTo3D(Vector3 *arrayOfPoints, unsigned int pointsCount, Vector
 		//					[y*x*C + z * s  y*y*C + c	    y*z*C - x * s]
 		//					[z*x*C - y * s  z*y*C + x * s	z*z*C + c])
 		Matrix rotationMatrix = Matrix();
-		rotationMatrix.m00 = axis.X * axis.X *C + c;					rotationMatrix.m01 = axis.X * axis.Y*C - axis.Z * s;			rotationMatrix.m02 = axis.X * axis.Z*C + axis.Y * s;
-		rotationMatrix.m10 = axis.Y * axis.X*C + axis.Z * s;			rotationMatrix.m11 = axis.Y * axis.Y*C + c;						rotationMatrix.m12 = axis.Y * axis.Z*C - axis.X * s;
-		rotationMatrix.m20 = axis.Z * axis.X*C - axis.Y * s;			rotationMatrix.m21 = axis.Z * axis.Y*C + axis.X * s;			rotationMatrix.m22 = axis.Z * axis.Z*C + c;
+		rotationMatrix.m00 = (double)axis.X * (double)axis.X * C + c;							rotationMatrix.m01 = (double)axis.X * (double)axis.Y*C - (double)axis.Z * s;			rotationMatrix.m02 = (double)axis.X * (double)axis.Z*C + (double)axis.Y * s;
+		rotationMatrix.m10 = (double)axis.Y * (double)axis.X * C + (double)axis.Z * s;			rotationMatrix.m11 = (double)axis.Y * (double)axis.Y*C + c;								rotationMatrix.m12 = (double)axis.Y * (double)axis.Z*C - (double)axis.X * s;
+		rotationMatrix.m20 = (double)axis.Z * (double)axis.X * C - (double)axis.Y * s;			rotationMatrix.m21 = (double)axis.Z * (double)axis.Y*C + (double)axis.X * s;			rotationMatrix.m22 = (double)axis.Z * (double)axis.Z*C + c;
 
 		for (size_t i = 0; i < pointsCount; i++)
 		{
